downloadVideo: video-payer-buttons.js -> download-video.js
